revalidacion de variables

+++ Operaciones.cs
@@ -3,7 +3,7 @@ namespace Operations
     internal class Create
     {
         //clase de validacion 0
-        public void validacion(long n)
+        public void validacion(double n)
         {
             if (n == 0)
             {
@@ -11,33 +11,33 @@ namespace Operations
             }
         }
         //componentes x and y de la fuerza ejercida
-        public double CompX(long F, long A)
+        public double CompX(double F, double A)
         {
             double Fx = F * Math.Cos(A * Math.PI / 180);
             return (Fx);
         }
-        public double CompY(long F, long A)
+        public double CompY(double F, double A)
         {
             double Fy = F * Math.Sin(A * Math.PI / 180);
             return (Fy);
         }
         //momentum
-        public double MomentoX(long dY, double Fx)
+        public double MomentoX(double dY, double Fx)
         {
             return (Fx * dY);
         }
-        public double MomentoY(long dX, double Fy)
+        public double MomentoY(double dX, double Fy)
         {
             return (Fy * dX);
         }
         //Calculo por medio de triangulo
-        public double ComponeteX(long Fh, long catad, long hip1)
+        public double ComponeteX(double Fh, double catad, double hip1)
         {
             //nota: no se puede hacer una divicion entre 2 longs
             double Fx = (Fh * ((double)catad / hip1));
             return (Fx);
         }
-        public double ComponenteY(long Fh, long catop, long hip1)
+        public double ComponenteY(double Fh, double catop, double hip1)
         {
             double Fy = (Fh * ((double)catop / hip1));
             return (Fy);
